+++ src/core/pre.js
@@ -2,4 +2,5 @@ const API = Module.API;
 const Hiwire = {};
 const Tests = {};
 API.tests = Tests;
+API.version = "0.21.2";
 Module.hiwire = Hiwire;