+++ src/tests/test_pyodide.py
@@ -1118,6 +1118,24 @@ def test_home_directory(selenium_standalone_noload):
     )
 
 
+def test_version_variable(selenium):
+    js_version = selenium.run_js(
+        """
+        return pyodide.version
+        """
+    )
+
+    core_version = selenium.run_js(
+        """
+        return pyodide._api.version
+        """
+    )
+
+    from pyodide import __version__ as py_version
+
+    assert js_version == py_version == core_version
+
+
 def test_sys_path0(selenium):
     selenium.run_js(
         """